pyproject.toml
[tool.poetry]
authors = ["Shuai Zhang <zhangshuai.ustc@gmail.com>"]
description = ""
name = "one"
version = "0.0.1"

[tool.poetry.dependencies]
absl-py = ">=0.12.0 <1.0.0"
python = "^3.6.1"
requests = "^2.25.1"
sh = "^1.14.1"

[tool.poetry.dev-dependencies]
cpplint = "^1.5.5"
isort = "^5.8.0"
pylint = "^2.7.4"
yapf = ">=0.31.0 <1.0.0"

[[tool.poetry.source]]
default = true
name = "tsinghua-mirror"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.isort]
force_single_line = true
indent = "    "
length_sort = true
line_length = 80
